complete reimplementation of Cadenza's lexer with Ragel
Ultimately this is a failure, at the time of committing the performance of
a pure-ruby Ragel based lexer is about 47% slower than the current StringScanner
implementation, which is already far too slow.

I'm commiting this work for future reference, perhaps in the next version of
Ragel the performance will be much improved.  For reference this version was
compiled with Ragel 6.7